test: reformat + convert die size to mm2 in verbose tests
